+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Show a notice when gps data is updated
 - Add an option to show more verbose information
+- Add default country support to remove geolocation
+
 
 ## [1.2] - 2023-05-01